wrapp app sidebar together
