+++ webClient/src/widgets/app-layout/AppSidebar.tsx
@@ -1,7 +1,17 @@
-import { Sidebar } from '@/shared/ui';
+import {
+    BookmarkCheckIcon,
+    BoxIcon,
+    FormIcon,
+    InboxIcon,
+    ListTodoIcon,
+} from 'lucide-react';
 
-import { AppSidebarHeader } from './AppSidebarHeader';
-import { AppSidebarNav } from './AppSidebarNav';
+import { Sidebar, SidebarHeader, NavGroup, Navigation } from '@/shared/ui';
+import { WorkspaceMenu } from './WorkspaceMenu';
+
+import type { TKey } from '@/shared/lib/i18n';
+
+import { routes } from '@/app/config/routes';
 
 type AppSidebarProps = {
     collapsed?: boolean;
@@ -14,8 +24,56 @@ export const AppSidebar = ({
 }: AppSidebarProps) => {
     return (
         <Sidebar>
-            <AppSidebarHeader collapsed={collapsed} onNavigate={onNavigate} />
-            <AppSidebarNav collapsed={collapsed} onNavigate={onNavigate} />
+            <SidebarHeader>
+                <WorkspaceMenu onNavigate={onNavigate} collapsed={collapsed} />
+            </SidebarHeader>
+            <Navigation>
+                {navGroups.map((navGroup) => (
+                    <NavGroup
+                        key={`navGroup-${navGroup.label ?? navGroup.items[0].to}`}
+                        navGroup={navGroup}
+                        collapsed={collapsed}
+                        onNavigate={onNavigate}
+                    />
+                ))}
+            </Navigation>
         </Sidebar>
     );
 };
+
+const navGroups = [
+    {
+        items: [
+            {
+                label: 'app.nav.inbox' as TKey,
+                to: routes.inbox(),
+                icon: InboxIcon,
+            },
+            {
+                label: 'app.nav.myIssues' as TKey,
+                to: routes.myIssues(),
+                icon: BookmarkCheckIcon,
+            },
+        ],
+    },
+    {
+        label: 'app.nav.workspace' as TKey,
+        items: [
+            {
+                label: 'app.nav.projects' as TKey,
+                to: routes.projects(),
+                icon: BoxIcon,
+            },
+            {
+                label: 'app.nav.issues' as TKey,
+                to: routes.issues(),
+                icon: ListTodoIcon,
+            },
+            {
+                label: 'app.nav.wiki' as TKey,
+                to: routes.wiki(),
+                icon: FormIcon,
+            },
+        ],
+    },
+];
